Fixes https://github.com/typelevel/cats-effect/issues/3929 by creating a custom `WeakList` datastructure inspired by the new `CallbackStack` introduced in https://github.com/typelevel/cats-effect/pull/3943.