@djspiewak 2024-02-23
## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   893.540 ±  1.695    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    70.238 ±  0.117    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    13.939 ±  0.124  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    48.600 ±  2.862  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2958.431 ± 11.388  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  2179.764 ± 20.288  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    50.254 ±  3.102  ops/min
```

## After

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score    Error    Units
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   903.096 ±  4.072    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    65.055 ±  5.206    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    13.677 ±  0.164  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    45.320 ±  2.022  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2965.083 ± 33.915  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  2194.897 ± 27.816  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    50.026 ±  3.685  ops/min
```

The `traverse` results are weird and probably just noise on my testing system. The rest is all within the margin of error.